Move more functions to utils
