+++ pumaguard/utils.py
@@ -2,7 +2,14 @@
 Some utility functions.
 """
 
+import glob
+import os
+import shutil
+
+import keras  # type: ignore
 import tensorflow as tf  # type: ignore
+
+from pumaguard.presets import Presets
 
 
 def initialize_tensorflow() -> tf.distribute.Strategy:
@@ -32,3 +39,94 @@ def initialize_tensorflow() -> tf.distribute.Strategy:
         print('WARNING: Not connected to TPU or GPU runtime; '
               'Will use CPU context')
         return tf.distribute.get_strategy()
+
+
+def copy_images(work_directory, lion_images, no_lion_images):
+    """
+    Copy images to work directory.
+    """
+    print(f'Copying images to working directory '
+          f'{os.path.realpath(work_directory)}')
+    for image in lion_images:
+        shutil.copy(image, f'{work_directory}/lion')
+    for image in no_lion_images:
+        shutil.copy(image, f'{work_directory}/no_lion')
+    print('Copied all images')
+
+
+def organize_data(presets: Presets, work_directory: str):
+    """
+    Organizes the data and splits it into training and validation datasets.
+    """
+    lion_images = []
+    for lion in presets.lion_directories:
+        lion_images += glob.glob(os.path.join(lion, '*JPG'))
+    no_lion_images = []
+    for no_lion in presets.no_lion_directories:
+        no_lion_images += glob.glob(os.path.join(no_lion, '*JPG'))
+
+    print(f'Found {len(lion_images)} images tagged as `lion`')
+    print(f'Found {len(no_lion_images)} images tagged as `no-lion`')
+    print(f'In total {len(lion_images) + len(no_lion_images)} images')
+
+    shutil.rmtree(work_directory, ignore_errors=True)
+    os.makedirs(f'{work_directory}/lion')
+    os.makedirs(f'{work_directory}/no_lion')
+
+    copy_images(work_directory=work_directory,
+                lion_images=lion_images,
+                no_lion_images=no_lion_images)
+
+
+def image_augmentation(image, with_augmentation: bool, augmentation_layers):
+    """
+    Use augmentation if `with_augmentation` is set to True
+    """
+    if with_augmentation:
+        for layer in augmentation_layers:
+            image = layer(image)
+    return image
+
+
+def create_datasets(presets: Presets, work_directory: str, color_mode: str):
+    """
+    Create the training and validation datasets.
+    """
+    # Define augmentation layers which are used in some of the runs
+    augmentation_layers = [
+        keras.layers.RandomFlip('horizontal'),
+        keras.layers.RandomRotation(0.01),
+        keras.layers.RandomZoom(0.05),
+        keras.layers.RandomBrightness((-0.1, 0.1)),
+        keras.layers.RandomContrast(0.1),
+        # keras.layers.RandomCrop(200, 200),
+        # keras.layers.Rescaling(1./255),
+    ]
+
+    # Create datasets(training, validation)
+    training_dataset, validation_dataset = \
+        keras.preprocessing.image_dataset_from_directory(
+            work_directory,
+            batch_size=presets.batch_size,
+            validation_split=0.2,
+            subset='both',
+            # Seed is always the same in order to ensure that we can reproduce
+            # the same training session
+            seed=123,
+            shuffle=True,
+            image_size=presets.image_dimensions,
+            color_mode=color_mode,
+        )
+
+    training_dataset = training_dataset.map(
+        lambda img, label: (image_augmentation(
+            image=img,
+            with_augmentation=presets.with_augmentation,
+            augmentation_layers=augmentation_layers), label),
+        num_parallel_calls=tf.data.AUTOTUNE,
+    )
+
+    training_dataset = training_dataset.prefetch(tf.data.AUTOTUNE)
+    validation_dataset = validation_dataset.prefetch(tf.data.AUTOTUNE)
+
+    return training_dataset, validation_dataset
